+++ src/modules/adsb-feeder/filesystem/root/opt/adsb/adsb-setup/utils/data.py
@@ -658,7 +658,8 @@ class Data:
         Env("FEED_ACARS_ACARSDRAMA", default=[False], tags=["feed_acars_acarsdrama", "is_enabled"]),
         # note that the ACARS and VDL Mode 2 strings start with a semicolon but HFDL does not -- see acars_router.yml
         Env("FEED_STRING_ACARS", default="", tags=["feed_string_acars"]),
-        Env("FEED_STRING_VDL2", default="", tags=["feed_string_vdl2"]),
+        Env("FEED_STRING_VDL2", default="", tags=["feed_string_vdl2_udp"]),
+        Env("FEED_STRING_VDL2_TCP", default="", tags=["feed_string_vdl2_tcp"]),
         Env("FEED_STRING_HFDL", default="", tags=["feed_string_hfdl"]),
         # AIS related stuff
         Env("AF_IS_SHIPFEEDER_ENABLED", default=False, tags=["shipfeeder", "is_enabled"]),